move project-avatar back into includes
