+++ src/presenters/includes/project-avatar.js
